Fix Navigation Drawer's text color. Originally it was black, now it's using the primary color.
The reason is that between material 1.2.1 and 1.8.0, the NavigationView constructor has changed. In particular textAppearanceSet was set to true, which didn't set up the itemTextColor, even though the textAppearance was 0. This was because the `app:itemTextAppearance="?attr/titleTextAppearance"` line in the XML was resolving to nothing, because the attribute is not meant to be used outside of a Toolbar.